// src/models/StrongExportRow.ts
export interface StrongExportRow {

    date: string;
    workoutName: string;
    exerciseName: string;
    setOrder: number;
    weight: number;
    weightUnit: string;
    reps: number;
    rpe: number;
    distance: number;
    distanceUnit: string
    seconds: number
    exerciseNotes: string;
    workoutNotes: string;
    workoutDuration: string;
}